+++ lib/shared_map.dart
@@ -2,3 +2,4 @@
 library shared_map;
 
 export 'src/shared_map_base.dart';
+export 'src/shared_map_cached.dart';