+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3
+
+- New `SharedMapCached`.
+- Fix `SharedStoreGeneric` and `SharedMapGeneric`.
+- Improve tests for `SharedStoreGeneric` and `SharedMapGeneric`.
+
 ## 1.0.2
 
 - `SharedMap`: